Remove kind=16 and make it host dependent
This is honestly bad idea but it makes things more feasible for x86.

+++ flang/test/Lower/Intrinsics/sign.f90
@@ -22,8 +22,8 @@ end subroutine
 
 ! CHECK-LABEL: sign_testr2
 subroutine sign_testr2(a, b, c)
-  real(KIND=16) a, b, c
+  real(KIND=8) a, b, c
   ! CHECK-NOT: fir.call @{{.*}}fabs
-  ! CHECK: fir.call @{{.*}}copysign{{.*}} : (f128, f128) -> f128
+  ! CHECK: fir.call @{{.*}}copysign{{.*}} : (f64, f64) -> f64
   c = sign(a, b)
 end subroutine
